+++ src/tools/file-switch_test.go
@@ -4,24 +4,20 @@ import (
 	"fmt"
 	"os"
 	"testing"
-	"path"
 
 	"axway.com/qlt-router/src/tools"
 	"github.com/esimov/gogu"
 )
 
-func CleanFiles(ctx, filename string) error {
-	extention := path.Ext(filename)
-	refname := filename[:len(filename)-len(extention)]
-
-	entries, err := tools.FileSwitchList(ctx, filename)
+func CleanFiles(ctx, filenamePrefix string, filenameSuffix string) error {
+	entries, err := tools.FileSwitchList(ctx, filenamePrefix, filenameSuffix)
 	if err != nil {
 		fmt.Println("FileSwitchList error ")
 		return err
 	}
 
 	for _, entry := range entries {
-		if entry[:len(refname)] != refname {
+		if entry[:len(filenamePrefix)] != filenamePrefix {
 			return err
 		}
 		err := os.Remove(entry)
@@ -35,19 +31,20 @@ func CleanFiles(ctx, filename string) error {
 func TestFileSwitch(t *testing.T) {
 	t.Parallel()
 
-	filename := "/tmp/testSwitch.txt"
+	prefix := "/tmp/testSwitch"
+	suffix := "txt"
 	ctx := "test"
 	data := "testdata"
 	n := 4
 
-	err := CleanFiles(ctx, filename)
+	err := CleanFiles(ctx, prefix, suffix)
 	if err != nil {
 		return
 	}
 
 	// real test
 	for i := 0; i < n+3; i++ {
-		entries, err := tools.FileSwitchList(ctx, filename)
+		entries, err := tools.FileSwitchList(ctx, prefix, suffix)
 		if err != nil {
 			t.Error("FileSwitchList error", err)
 			return
@@ -55,13 +52,13 @@ func TestFileSwitch(t *testing.T) {
 		expectedLength := gogu.Min(i, n)
 
 		if len(entries) != expectedLength {
-			t.Error(fmt.Sprint("It", i) + " expecting number of file " + fmt.Sprint(expectedLength, " got ", len(entries), " : ", entries))
+			t.Error(fmt.Sprint("It ", i) + " expecting number of file " + fmt.Sprint(expectedLength, " got ", len(entries), " : ", entries))
 			return
 		}
 
-		fullfilename, err := tools.FileSwitch(ctx, filename, n)
+		fullfilename, err := tools.FileSwitch(ctx, prefix, suffix, n)
 		if err != nil {
-			t.Error("unexpected file rotate", err)
+			t.Error("unexpected file switch", err)
 			return
 		}
 
@@ -72,7 +69,7 @@ func TestFileSwitch(t *testing.T) {
 		}
 	}
 
-	err = CleanFiles(ctx, filename)
+	err = CleanFiles(ctx, prefix, suffix)
 	if err != nil {
 		return
 	}