+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 # 0.2.0 (dev)
 - BREAKING-CHANGE : refactor streams to add `reader`, `transforms`, `writer` section
 - BREAKING-CHANGE : connector now are identified through `type` key instead of `name`
+- BREAKING-CHANGE : `file-reader` and `file-writer` now use `filenamePrefix` and `filenameSuffix` instead of `filename`
 - use extension `.ser.yml` to enable schema validation from `qlt-router-schema.yml` (under `vscode`)
+- `file-writer` parameter `maxSize` is now in MByte
+
 
 # 0.0.1sink
 - connector: `file`, add `MaxFile`, `MaxSize` for automatic file rotation and jsonfile sample